add basic query restriction

+++ routes/days.js
@@ -7,7 +7,9 @@ const Day = require('../models/Days.js')
 let currentTime = new Date().toLocaleString()
 
 router.get('/test', (req, res) => {
-  res.send(`Day test route is running on /api/v1/days/test at ${currentTime}`)
+  res.send(
+    `Day test route is running on /api/v1/days/test at ${currentTime}, ${req.query.data}`
+  )
 })
 
 // @desc:   Get all days information
@@ -21,18 +23,21 @@ router.get('/', (req, res) => {
 // @desc:   Create new day
 // @route:  POST /add
 router.post('/add', (req, res) => {
-  const newDay = new Day({
-    day: req.body.day,
-    seasonIn: req.body.seasonIn,
-    monthIn: req.body.monthIn,
-    nDay: req.body.nDay,
-    holiday: req.body.holiday,
-    birthday: req.body.birthday,
-  })
-  newDay
-    .save()
-    .then(() => res.json(newDay))
-    .catch((err) => res.status(400).json(`Error ${err}`))
+  console.log(req.query.data)
+  if (req.query.data == process.env.DATA) {
+    const newDay = new Day({
+      day: req.body.day,
+      seasonIn: req.body.seasonIn,
+      monthIn: req.body.monthIn,
+      nDay: req.body.nDay,
+      holiday: req.body.holiday,
+      birthday: req.body.birthday,
+    })
+    newDay
+      .save()
+      .then(() => res.json(newDay))
+      .catch((err) => res.status(400).json(`Error ${err}`))
+  }
 })
 
 // @desc:   Get day by id
@@ -46,27 +51,31 @@ router.get('/:id', (req, res) => {
 // @desc:   Update day
 // @route:  PUT /:id
 router.patch('/:id', (req, res) => {
-  Day.findByIdAndUpdate(req.params.id)
-    .then((day) => {
-      ;(day.dayOYear = req.body.dayOYear),
-        (day.seasonIn = req.body.seasonIn),
-        (day.monthIn = req.body.monthIn),
-        (day.newDate = req.body.newDate),
-        (day.holiday = req.body.holiday)
-      day
-        .save()
-        .then(() => res.json(`Day had been updated.`))
-        .catch((err) => res.status(400).json(`Error: ${err}`))
-    })
-    .catch((err) => res.status(400).json(`Error: ${err}`))
+  if (req.query.data == process.env.DATA) {
+    Day.findByIdAndUpdate(req.params.id)
+      .then((day) => {
+        ;(day.dayOYear = req.body.dayOYear),
+          (day.seasonIn = req.body.seasonIn),
+          (day.monthIn = req.body.monthIn),
+          (day.newDate = req.body.newDate),
+          (day.holiday = req.body.holiday)
+        day
+          .save()
+          .then(() => res.json(`Day had been updated.`))
+          .catch((err) => res.status(400).json(`Error: ${err}`))
+      })
+      .catch((err) => res.status(400).json(`Error: ${err}`))
+  }
 })
 
 // @desc:   Delete day
 // @route:  DELETE /:id
 router.delete('/:id', (req, res) => {
-  Day.findOneAndDelete(req.params.id)
-    .then(() => res.json(`The day with id ${req.params.id} has been delted`))
-    .catch((err) => res.status(400).json(`Error: ${err}`))
+  if (req.query.data == process.env.DATA) {
+    Day.findOneAndDelete(req.params.id)
+      .then(() => res.json(`The day with id ${req.params.id} has been deleted`))
+      .catch((err) => res.status(400).json(`Error: ${err}`))
+  }
 })
 
 module.exports = router
